tiny parser module, file change logic server side
added the parser module for shared functions, wasn't sure what to name
it yet so I just named it after the singular shared function. Added the
server side logic and I moved the qrexec module into an external crate
because Its pretty generic code in that It could be useful for a variety
of applications, also I think I can speed it up a lot if I read some of
the qrexec code from the qubes folks and see how its working with the
xen vchan code.