err msg to shared_consts.rs, book_path notify def
book directory inotify handling defined in client.rs client_main, this
ends up querying for a book from the server whenever the client opens a
book file up. The book files are originally loaded onto the client vm
without the book contents because that could be a rather large amount of
data. Zathura features hot reloading of files so the editor should just
update in place once the book loads in from the server.