added send_sfile_tree fn
send_sfile_tree fn is the function used to respond to a clients request
for the server to send over the full sfile tree. It is still incomplete
there is an incongruency with the protocols actual definition and it's
late in the night so I will figure it out tommorow.