loads booknames periodically, non_blocking inotify
booknames need to be reloaded when the user adds a new book to the
library on the server VM. Inotify can't block on reads because
that will introduce potentially indefinite latency while the user
waits for a file to load into their books directory.
Everything looks good to me right now, there will indoubtably be
some edge cases and bugs to fix though.